Metrics test and typing
