+++ mapie/utils.py
@@ -8,11 +8,11 @@ from sklearn.linear_model import LogisticRegression
 from sklearn.pipeline import Pipeline
 from sklearn.model_selection import BaseCrossValidator, KFold, LeaveOneOut
 from sklearn.utils.validation import (
-    _check_sample_weight, _num_features, check_is_fitted
+    _check_sample_weight, _num_features, check_is_fitted, column_or_1d
 )
 from sklearn.utils import _safe_indexing
 from sklearn.model_selection import train_test_split
-
+from sklearn.utils.multiclass import type_of_target
 
 from ._compatibility import np_quantile
 from .conformity_scores import AbsoluteConformityScore, ConformityScore
@@ -20,7 +20,7 @@ from ._typing import ArrayLike, NDArray
 
 
 def check_null_weight(
-    sample_weight: Optional[ArrayLike], X: ArrayLike, y: ArrayLike
+    sample_weight: Optional[NDArray], X: ArrayLike, y: ArrayLike
 ) -> Tuple[Optional[NDArray], ArrayLike, ArrayLike]:
     """
     Check sample weights and remove samples with null sample weights.
@@ -70,7 +70,6 @@ def check_null_weight(
         X = _safe_indexing(X, non_null_weight)
         y = _safe_indexing(y, non_null_weight)
         sample_weight = _safe_indexing(sample_weight, non_null_weight)
-    sample_weight = cast(Optional[NDArray], sample_weight)
     return sample_weight, X, y
 
 
@@ -78,7 +77,7 @@ def fit_estimator(
     estimator: Union[RegressorMixin, ClassifierMixin],
     X: ArrayLike,
     y: ArrayLike,
-    sample_weight: Optional[ArrayLike] = None,
+    sample_weight: Optional[NDArray] = None,
 ) -> Union[RegressorMixin, ClassifierMixin]:
     """
     Fit an estimator on training data by distinguishing two cases:
@@ -669,16 +668,17 @@ def compute_quantiles(vector: NDArray, alpha: NDArray) -> NDArray:
 def check_calib_set(
     X: ArrayLike,
     y: ArrayLike,
-    sample_weight: Optional[ArrayLike] = None,
+    sw: Optional[NDArray] = None,
     X_calib: Optional[ArrayLike] = None,
     y_calib: Optional[ArrayLike] = None,
-    sample_weight_calib: Optional[ArrayLike] = None,
+    sw_calib: Optional[NDArray] = None,
     calib_size: Optional[float] = 0.3,
     random_state: Optional[Union[int, np.random.RandomState]] = None,
     shuffle: Optional[bool] = True,
     stratify: Optional[ArrayLike] = None,
 ) -> Tuple[
-    ArrayLike, ArrayLike, ArrayLike, ArrayLike, Optional[ArrayLike]
+    ArrayLike, ArrayLike, ArrayLike, ArrayLike,
+    Optional[NDArray], Optional[NDArray]
 ]:
     """
     Check if a calibration set has already been defined, if not, then
@@ -706,12 +706,12 @@ def check_calib_set(
 
     """
     if X_calib is None or y_calib is None:
-        if sample_weight_calib is not None:
+        if sw_calib is not None:
             warnings.warn(
-            "WARNING: sample weight for calibration"
-            + " were provided without X_calib and y_calib."
-        )
-        if sample_weight is None:
+                "WARNING: sample weight for calibration"
+                + " were provided without X_calib and y_calib."
+            )
+        if sw is None:
             (
                 X_train, X_calib, y_train, y_calib
             ) = train_test_split(
@@ -722,30 +722,28 @@ def check_calib_set(
                     shuffle=shuffle,
                     stratify=stratify
             )
-            sample_weight_train = sample_weight
+            sw_train = sw
         else:
             (
                     X_train,
                     X_calib,
                     y_train,
                     y_calib,
-                    sample_weight_train,
-                    sample_weight_calib,
+                    sw_train,
+                    sw_calib,
             ) = train_test_split(
                     X,
                     y,
-                    sample_weight,
+                    sw,
                     test_size=calib_size,
                     random_state=random_state,
                     shuffle=shuffle,
                     stratify=stratify
             )
     else:
-        X_train, y_train, sample_weight_train = X, y, sample_weight
+        X_train, y_train, sw_train = X, y, sw
     X_train, X_calib = cast(ArrayLike, X_train), cast(ArrayLike, X_calib)
     y_train, y_calib = cast(ArrayLike, y_train), cast(ArrayLike, y_calib)
-    sw_train = cast(ArrayLike, sample_weight_train)
-    sw_calib = cast(ArrayLike, sample_weight_calib)
     return X_train, y_train, X_calib, y_calib, sw_train, sw_calib
 
 
@@ -806,3 +804,127 @@ def check_estimator_classification(
                 "'classes_' attribute."
             )
     return estimator
+
+
+def get_binning_groups(
+    y_score: NDArray,
+    num_bins: int,
+    strategy: str,
+) -> Union[NDArray, NDArray]:
+    """_summary_
+    Parameters
+    ----------
+    y_score : _type_
+        The scores given from the calibrator.
+    num_bins : _type_
+        Number of bins to make the split in the y_score.
+    strategy : _type_
+        The way of splitting the predictions into different bins.
+    Returns
+    -------
+    _type_
+        Returns the upper and lower bound values for the bins and the indices
+        of the y_score that belong to each bins.
+    """
+    if strategy == "quantile":
+        quantiles = np.linspace(0, 1, num_bins)
+        bins = np.percentile(y_score, quantiles * 100)
+    elif strategy == "uniform":
+        bins = np.linspace(0.0, 1.0, num_bins)
+    elif strategy == "array split":
+        bin_groups = np.array_split(y_score, num_bins)
+        bins = np.sort(
+            np.array(
+                [bin_group.max() for bin_group in bin_groups[:-1]]+[np.inf]
+                )
+        )
+    else:
+        ValueError("We don't have this strategy")
+    bin_assignments = np.digitize(y_score, bins, right=True)
+    return bins, bin_assignments
+
+
+def calc_bins(
+    y_score: NDArray,
+    y_true: NDArray,
+    num_bins: int,
+    strategy: str,
+) -> Union[NDArray, NDArray, NDArray, NDArray]:
+    """
+    For each bins, calculate the accuracy, average confidence and size.
+    Parameters
+    ----------
+    y_score : _type_
+        The scores given from the calibrator.
+    y_true : _type_
+        The "true" values, target for the calibrator.
+    num_bins : _type_
+        Number of bins to make the split in the y_score.
+    strategy : _type_
+        The way of splitting the predictions into different bins.
+    Returns
+    -------
+    _type_
+        Multiple arrays, the upper and lower bound of each bins,
+        indices of y that belong to each bins, the accuracy,
+        confidence and size of each bins.
+    """
+    bins, binned = get_binning_groups(y_score, num_bins, strategy)
+    bin_accs = np.zeros(num_bins)
+    bin_confs = np.zeros(num_bins)
+    bin_sizes = np.zeros(num_bins)
+
+    for bin in range(num_bins):
+        bin_sizes[bin] = len(y_score[binned == bin])
+        if bin_sizes[bin] > 0:
+            bin_accs[bin] = np.divide(
+                np.sum(y_true[binned == bin]),
+                bin_sizes[bin],
+            )
+            bin_confs[bin] = np.divide(
+                np.sum(y_score[binned == bin]),
+                bin_sizes[bin],
+            )
+    return bins, bin_accs, bin_confs, bin_sizes
+
+
+def check_split_strategy(
+    strategy: Optional[str]
+) -> str:
+    valid_split_strategies = ["uniform", "quantile", "array split"]
+    if strategy is None:
+        strategy = "uniform"
+    if strategy not in valid_split_strategies:
+        raise ValueError(
+            "Please provide a valid splitting strategy."
+        )
+    return strategy
+
+
+def check_number_bins(
+    num_bins: int
+) -> None:
+    if isinstance(num_bins, int) is False:
+        raise ValueError(
+            "Please provide a bin number as an integer."
+        )
+    elif num_bins < 1:
+        raise ValueError(
+            """
+            Please provide a bin number greater than
+            or equal to  1.
+            """
+        )
+
+
+def check_binary_zero_one(
+    y_true: ArrayLike
+) -> NDArray:
+    cast(NDArray, column_or_1d(y_true))
+    assert type_of_target(y_true) == "binary"
+    if (np.unique(y_true) != np.array([0, 1])).all():
+        idx_min = np.where(y_true == np.min(y_true))[0]
+        y_true[idx_min] = 0
+        idx_max = np.where(y_true == np.max(y_true))[0]
+        y_true[idx_max] = 1
+    return y_true
